# docs/tarot-api-spec.yaml
openapi: 3.0.3
info:
  title: Mystic Arcana Tarot API
  description: |
    Complete REST API for tarot card reading functionality, including deck management,
    card drawing, reading persistence, and retrieval capabilities.
    
    Built with production-grade error handling, validation, and comprehensive logging.
  version: 1.0.0
  contact:
    name: Mystic Arcana API Support
    url: https://mysticarcana.dev/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://mysticarcana.vercel.app/api
    description: Production server
paths:
...
  /tarot/shuffle:
    post:
      summary: Shuffle tarot deck
      description: |
        Shuffle a tarot deck using various algorithms and return shuffle state information.
        Includes entropy calculation and optional deck preview.
      operationId: shuffleDeck
      tags:
        - Tarot Cards
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                deckId:
                  type: string
                  format: uuid
                  default: "00000000-0000-0000-0000-000000000001"
                  description: Deck UUID to shuffle
                algorithm:
                  type: string
                  enum: [fisher-yates, riffle, overhand]
                  default: fisher-yates
                  description: Shuffle algorithm
                includePreview:
                  type: boolean
                  default: false
                  description: Include sample cards in response
                userId:
                  type: string
                  format: uuid
                  description: User ID for user-specific shuffles
            examples:
              basic-shuffle:
                summary: Basic Fisher-Yates shuffle
                value:
                  algorithm: "fisher-yates"
              riffle-shuffle-with-preview:
                summary: Riffle shuffle with preview
                value:
                  algorithm: "riffle"
                  includePreview: true
      responses:
        '200':
          description: Deck shuffled successfully
          headers:
            X-Shuffle-Time:
              description: Time taken to shuffle (ms)
              schema:
                type: string
            X-Shuffle-ID:
              description: Unique identifier for this shuffle
              schema:
                type: string
            X-Entropy:
              description: Calculated shuffle entropy
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShuffleResponse'
        '404':
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tarot/save-reading:
    post:
      summary: Save tarot reading
      description: |
        Persist a completed tarot reading to the database with full metadata,
        interpretation, and optional user notes.
      operationId: saveReading
      tags:
        - Tarot Readings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveReadingRequest'
            examples:
              three-card-reading:
                summary: Three card reading save
                value:
                  userId: "user-123"
                  spreadType: "three-card"
                  cards:
                    - id: "the-fool"
                      name: "The Fool"
                      position: "Past"
                      isReversed: false
                      meaning: "New beginnings, innocence, spontaneity"
                    - id: "the-magician"
                      name: "The Magician"
                      position: "Present"
                      isReversed: false
                      meaning: "Manifestation, resourcefulness, power"
                    - id: "the-world"
                      name: "The World"
                      position: "Future"
                      isReversed: true
                      meaning: "Seeking personal closure, incomplete goals"
                  interpretation: "Your journey shows a progression from new beginnings to mastery, with final completion requiring inner work."
                  question: "What does my career path look like?"
                  tags: ["career", "guidance"]
      responses:
        '201':
          description: Reading saved successfully
          headers:
            X-Save-Time:
              description: Time taken to save reading (ms)
              schema:
                type: string
            X-Reading-ID:
              description: ID of the saved reading
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveReadingResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get reading save statistics
      description: Get user reading statistics and save limits
      operationId: getReadingStats
      tags:
        - Tarot Readings
      parameters:
        - name: userId
          in: query
          required: true
          description: User ID to get statistics for
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingStatsResponse'
...
components:
  schemas:
...
    ShuffleResponse:
      type: object
      properties:
        success:
          type: boolean
        shuffleId:
          type: string
          description: Unique shuffle identifier
        deckId:
          type: string
          description: Deck that was shuffled
        cardCount:
          type: integer
          description: Number of cards in deck
        shuffleState:
          type: object
          properties:
            algorithm:
              type: string
              description: Algorithm used
            timestamp:
              type: string
              format: date-time
            entropy:
              type: number
              description: Calculated randomness measure
        preview:
          type: object
          properties:
            topCard:
              type: object
            middleCard:
              type: object
            bottomCard:
              type: object
          description: Optional card preview
    SaveReadingRequest:
      type: object
      required:
        - userId
        - spreadType
        - cards
        - interpretation
      properties:
        userId:
          type: string
          format: uuid
        spreadType:
          type: string
          enum: [single, three-card, celtic-cross]
        cards:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              position:
                type: string
              isReversed:
                type: boolean
              meaning:
                type: string
              frontImage:
                type: string
              backImage:
                type: string
        interpretation:
          type: string
        question:
          type: string
        notes:
          type: string
        cosmicInfluence:
          type: object
        drawId:
          type: string
        isPublic:
          type: boolean
          default: false
        tags:
          type: array
          items:
            type: string
    SaveReadingResponse:
      type: object
      properties:
        success:
          type: boolean
        readingId:
          type: string
          format: uuid
        savedAt:
          type: string
          format: date-time
    ReadingStatsResponse:
      type: object
      properties:
        success:
          type: boolean
        userId:
          type: string
        statistics:
          type: object
          properties:
            totalReadings:
              type: integer
            readingsByType:
              type: object
              properties:
                single:
                  type: integer
                three-card:
                  type: integer
                celtic-cross:
                  type: integer
            lastReading:
              type: string
              format: date-time
              nullable: true
            limits:
              type: object
              properties:
                maxReadingsPerDay:
                  type: integer
                maxReadingsPerMonth:
                  type: integer
                storageRetention:
                  type: string
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        details:
          type: string
          description: Additional error details (optional)
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...